[MERGE] Feat/board 기능

+++ controller/board/deletePost.js
@@ -6,10 +6,20 @@ const deletePost = async (req, res, next) => {
   try {
     const postId = Number(req.params.post_id);
 
-    await models.Comment.destroy({ where: { postId: postId } });
-    await models.Post.destroy({ where: { id: postId } });
+    const Post = await models.Post.findOne({
+      where: { id: postId }
+    });
 
-    return res.status(200).send('DELETE_POST_SUCCESS');
+    if (!Post) {
+      return res.status(400).send('POST_DOESNT_EXIST');
+    } else if (req.decoded.id != Post.userId || req.decoded.isAdmin != true) {
+      return res.status(401).send('NO_PERMISSIONS');
+    } else {
+      await models.Comment.destroy({ where: { postId: postId } });
+      await models.Post.destroy({ where: { id: postId } });
+
+      return res.status(200).send('DELETE_POST_SUCCESS');
+    }
   } catch (err) {
     req.message = 'DELETE_POST';
     next(err);
